Avoid parsing (and caching) certificates from keyless.
Thanks to google/boringssl@3a2b47a ("Don't use |X509|
objects in |CERT|, by default."), BoringSSL now stores
certificates as binary blobs, rather than in parsed form.

This means we can avoid parsing our certificates entirely
and use the raw DER encoded certificates returned from
keyless directly. This also means we can entirely remove
the certificate cache, which was complicated and untested.

This requires a patch for BoringSSL to allow raw certificates
to be passed in directly. google/boringssl@3a2b47a did not add
an API to pass raw certificates, but the required changes are
minimal.

|ssl_cert_parse_pubkey| and |ssl_cert_skip_to_spki| are pulled
from the BoringSSL tip (google/boringssl@f71036e) to allow the
public key to be parsed from the leaf certificate without
parsing the entire certificate. They have been merged into a
single function. Extracting the public key is needed to for
|ngx_http_keyless_key_type| and
|ngx_http_keyless_key_max_signature_len|.

Because the certificates are no longer parsed, the logic in
|ngx_ssl_session_id_context| is no good. Instead a new session
id context is generated from a truncated SHA256 hash of the
octet string "keyless-nginx" concatenated with the payload.
This means the serialised sessions are no longer compatible
with non-keyless sessions.

This is a substantial change, but it greatly simplifies
keyless-nginx and reduces the code size.